rename mos6502 backend to m6502
